# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravue API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Gerar token de autenticão",
                "operationId": "22e3905726dd342e825c772ea8acb378",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "device": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "402": {
                        "description": "Validação de dados"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                }
            }
        },
...
        "/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Consultar usuário autenticado",
                "operationId": "975a2eebe95c72f232da26e3df07b6de",
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/password/reset": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Realizar alteração de senha",
                "operationId": "c9a37ed7e24bb8c8e5c8afb8e3368cf9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "code": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "400": {
                        "description": "Falha na requisição"
                    },
                    "402": {
                        "description": "Validação de dados"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                }
            }
        },
...
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrar conta de usuário",
                "operationId": "405b3491db441a409eaa828dde286aaa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "business": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "device": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/auth/password/code": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Enviar código de recuperação de senha",
                "operationId": "0ef183242819a285c9a71ae7ea68b4fc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "402": {
                        "description": "Validação de dados"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                }
            }
        },
        "/auth/password/verify": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Validar código de recuperação",
                "operationId": "3a356d4a483c52e43f1f83649f889c5b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "code": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "400": {
                        "description": "Falha na requisição"
                    },
                    "402": {
                        "description": "Validação de dados"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}